requirements.txt
streamlit
pymupdf
openai
numpy
faiss-cpu
sentence-transformers
scikit-learn